backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
Flask-JWT-Extended==4.6.0
flask-mongoengine==1.0.0
Werkzeug==3.0.3
requests==2.32.3
python-dotenv==1.0.1
flask-socketio==5.3.6
eventlet==0.36.1
gunicorn==22.0.0
redis==5.0.4
flask-limiter==3.7.0
flask-caching==2.1.0
python-socketio==5.11.2
bidict==0.23.1
